requirements.txt
requests~=2.26.0
aiogram~=2.14.3
Pillow~=8.3.1
beautifulsoup4~=4.9.3
mysql-connector-python~=8.0.26
envparse~=0.2.0
python-dotenv~=0.18.0
aiohttp~=3.7.4.post0